Implement retransmission of packet on peer request
As specified by RFC 1350, in the case where a waiting receiver
peer resends his last packet, this must be taken as a request
to the sender to retransmit also his last packet.

Hence duplicate receive of either DAT of ACK packets must be
implemented by resending the own last packet iif the duplicate
DAT or ACK is for the previous block.

Implement it as stated above when the last block.
For older handled block, ignore the packet.

Fix issues with observed clients (resp. server) implementing
restransmission request as per the protocol, in which case
the client (resp. server) may get stuck or exhaust his retry
count and fail.

Observed for instance with u-boot tftp client when the u-boot
client timeout is smaller than the tftpy server timeout.