Accept None as TFTP root
If None is given as tftproot upon tftpServer creation, the server always
forwards the file returned by its dyn_file_func method. This is useful if full
access to the file system should not be granted for the tftp client. The
client can still get files, but is restricted as appropriate by the
dyn_file_func method.